refactor(runtime): split skills.ts along discovery/metadata/context/state seams
Split the 1921-line skills.ts into focused sub-modules, keeping all
public exports stable through a re-export barrel and index.ts updates:

- skills-discovery.ts: resolveSkillDiscoveryPaths, normalizeSkillSource,
  scanSkills* family, scanSkillDir, ScannedSkill/RuntimeSkillDefinition types
- skills-metadata.ts: validateSkillMetadata, parseSkillFrontMatter,
  manifest-parsing family, SkillManifest/validation types, cleanPromptText
- skills-context.ts: gateSkillsByHostCapabilities, selectSkillsForContext,
  searchSkills + scoring, prompt rendering, loadSkillInstructions
- skills-state.ts: readSkillRuntimeState, writeSkillRuntimeState/Preferences
- skills-agent-tools.ts: buildSkillAgentTool, buildSkillSearchAgentTool,
  SkillShadowSelectionTracker
- path-containment.ts: moved readContainedRegularFile,
  readContainedRegularTextFile, writeContainedRegularTextFile, isRecord
  (generic infrastructure already imported cross-package by desktop)

skills.ts is now a thin re-export barrel so existing 'from ./skills.js'
imports continue to work. Internal consumers (skill-invocation.ts,
tool-catalog-derive.ts) updated to import from specific sub-modules.

Closes #1408
